refactor: split working_repository tests
